Used AI to implement the full API surface of libplctag, fix bugs and add a test suite.

+++ lib/data_types/lint_plc_mapper.dart
@@ -3,7 +3,7 @@ import 'package:libplctag_dart/tag.dart';
 
 class LintPlcMapper extends PlcMapperBase<int> {
   @override
-  int decodeAtOffset(Tag tag, int offset) => tag.getInt32(offset);
+  int decodeAtOffset(Tag tag, int offset) => tag.getInt64(offset);
 
   @override
   int? get elementSize => 8;
